refactor: ApiResponse 파일 통합
